+++ src/mobius/Components.cpp
@@ -3360,7 +3360,7 @@ PUBLIC void FocusButton::mousePressed(MouseEvent* e)
     a->setTargetTrack(mTrack);
 
     // Trigger id will be the address of the component
-    a->id = (long)this;
+    a->id = (uintptr_t)this;
     a->trigger = TriggerUI;
     a->triggerMode = TriggerModeOnce;
     
@@ -3633,7 +3633,7 @@ PUBLIC void TrackGroupButton::mousePressed(MouseEvent* e)
     a->setTargetTrack(mTrack);
 
     // Trigger id will be the address of the component
-    a->id = (long)this;
+    a->id = (uintptr_t)this;
     a->trigger = TriggerUI;
     a->triggerMode = TriggerModeOnce;
 
@@ -5282,7 +5282,7 @@ LoopStack::LoopStack(MobiusInterface* m, int track)
     mAction->setFunction(LoopN);
     mAction->setTargetTrack(track);
     // Trigger id will be the address of the component
-    mAction->id = (long)this;
+    mAction->id = (uintptr_t)this;
     mAction->trigger = TriggerUI;
     // we're not passing down up transitions of the mouse button
     mAction->triggerMode = TriggerModeOnce;
